move ssl stuff to its own file to better test it
