+++ lib/control.js
@@ -2,11 +2,11 @@
 
 var uuid = require('node-uuid');
 var path = require('path');
+var ssl = require('./ssl');
 var fs = require('fs-extra');
 var mkdirp = require('mkdirp');
 var childProcess = require('child_process');
 var isRelative = require('./util').isRelative;
-var os = require('os');
 var configGenerator = require('./config_generator');
 
 exports.start = function(config) {
@@ -60,7 +60,7 @@ exports.start = function(config) {
         path.join(tmpDir, 'server.key')
       );
     } else {
-      createCertificates(tmpDir);
+      ssl.createCertificates(tmpDir);
     }
   }
 
@@ -87,41 +87,4 @@ exports.stop = function() {
   childProcess.spawnSync('pkill', ['-f', 'nginx: master process']);
 };
 
-function createCertificates(tmpDir) {
-  var keyPath = path.join(tmpDir, 'server.key');
-  var hostname = os.hostname();
 
-  execAndCheck(
-    'openssl genrsa -des3 -out ' + keyPath + ' -passout pass:temp 2048',
-    true
-  );
-
-  var signingRequestPath = path.join(tmpDir, 'server.csr');
-  execAndCheck(
-    'openssl req -new -key ' + keyPath + ' -out ' + signingRequestPath +
-      ' -passin pass:temp -batch -subj "/OU=' + hostname + '"',
-    true
-  );
-
-  // remove the password
-  execAndCheck(
-    'openssl rsa -in ' + keyPath + ' -out ' + keyPath + ' -passin pass:temp',
-    true
-  );
-
-  var certificatePath = path.join(tmpDir, 'server.crt');
-  execAndCheck(
-    'openssl x509 -req -days 365 -in ' + signingRequestPath + ' -signkey ' +
-      keyPath + ' -out ' + certificatePath,
-    true
-  );
-}
-
-function execAndCheck(command, silent) {
-  // passing an empty array for option stdio results in silencing the command,
-  // that is, do not echo stdout/stderr of the child process
-  var options = silent ? { stdio: [] } : {};
-  childProcess.execSync(command, options);
-}
-
-exports.createCertificates = createCertificates;
